# Conflicts:
#   backend/sikdorang/guide/urls.py

+++ backend/sikdorang/guide/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import get_object_or_404
 from django.contrib.auth import get_user_model
 from rest_framework.decorators import api_view, authentication_classes
 from django.contrib.auth import get_user_model
-from django.http import HttpResponse
+from django.http import HttpResponse, JsonResponse
 from rest_framework.response import Response
 # from rest_framework import viewsets
 from .serializers import *
@@ -43,16 +43,25 @@ def list_tour(request):
 
 @api_view(['GET'])
 def detail_tour(request, tour_pk):
-    tour = TripItemModel.objects.filter(pk=tour_pk)
-    serializer = TourDetailSerializer(tour, many=True)
-    return Response(serializer.data)
-
+    User = get_user_model()
+    user = get_object_or_404(User, pk=request.user.pk)
+    tours = TripItemModel.objects.filter(pk=tour_pk)[0]
+    paiders = GuideTour.objects.filter(trip_item=tour_pk)
+    flag = False
+    for paider in paiders:
+        if paider.user == user:
+            flag = True
+    serializer = TourDetailSerializer(tours)
+    return JsonResponse({"result": serializer.data, "flag": flag})
+ 
 
 @api_view(['GET'])
 def list_guide(request):
     User = get_user_model()
     user = get_object_or_404(User, pk=request.user.pk)
-    trips = TripItemModel.objects.filter(user=user)
+    now = datetime.datetime.now()
+    nowDate = now.strftime('%Y%m%d')
+    trips = TripItemModel.objects.filter(start_date__gte=int(nowDate), user=user).order_by('start_date')
     serializer = GuideSerializer(trips, many=True)
     return Response(serializer.data)
 
@@ -74,8 +83,8 @@ def paid(request, trip_pk):
 def paidtour(request):
     User = get_user_model()
     user = get_object_or_404(User, pk=request.user.pk)
-    tours = TripItemModel.objects.filter(user=user)
-    serializer = TourSerializer(tours, many=True)
+    tours = GuideTour.objects.filter(user=user)
+    serializer = PaidSerializer(tours, many=True)
     return Response(serializer.data)
 
 @api_view(['GET'])
@@ -84,14 +93,32 @@ def paider(request, trip_pk):
     serializer = PaidSerializer(paider, many=True)
     return Response(serializer.data)
 
+# @api_view(['PUT'])
+# def update_guide(request, trip_pk):
+#     User = get_user_model()
+#     user = get_object_or_404(User, pk=request.user.pk)
+#     tour = get_object_or_404(TripItemModel, pk=trip_pk)
+#     if tour.user == user:
+#         serializer = GuideItemSerializer(data=request.data)
+#         if serializer.is_valid():
+#             serializer.update(tour, request.data)
+#             return Response(serializer.data)
+        
+#     return HttpResponse('Something Wrong')
+
 @api_view(['DELETE'])
-def delete_tour(request, tour_pk):
+def delete_guide(request, trip_pk):
     User = get_user_model()
     user = get_object_or_404(User, pk=request.user.pk)
-    tour = TripItemModel.objects.filter(user=user, id=tour_pk)
+    tour = get_object_or_404(TripItemModel, pk=trip_pk)
+    
     if tour.user == user:
-        tour.delete()
-        return HttpResponse('잘 지워짐')
+        paider = GuideTour.objects.filter(trip_item=trip_pk)
+        if paider  :
+            return HttpResponse('이미 결제한 사람 있어서 삭제 불가')
+        else :
+            tour.delete()
+            return HttpResponse('잘 지워짐')
     return HttpResponse('니 글 아님 ㅅㄱ')
 
 @api_view(['PUT'])
